+++ lib/google_maps_place_picker.dart
@@ -3,4 +3,5 @@ library google_maps_place_picker_mb;
 export 'src/models/pick_result.dart';
 export 'src/components/floating_card.dart';
 export 'src/components/rounded_frame.dart';
+export 'src/models/circle_area.dart';
 export 'src/place_picker.dart';